requirements.txt
pandas
dash==2.14.1
plotly==5.18.0
pymongo==4.6.1
gunicorn==21.2.0
python-dotenv==1.0.0
dash_bootstrap_components==1.5.0
phonenumbers==8.13.29
pycountry==22.3.5
